Reduce the loop device queue depth
The default queue depth for a loop device is 128 while the queue depth
for the storage device in my Android test setup is 32. This results in
high latencies. A common approach for reducing block device latency is
by reducing the queue depth. Hence this patch that reduces the queue
depth for loop devices.

Bug: 194726804
Test: Built Android images, installed these on a test device and verified the queue depth.
Depends-On: 1850656
Change-Id: Ic0b3894a66c3312b038a52072f91e5fa9b6dcdc0
Signed-off-by: Bart Van Assche <bvanassche@google.com>